move spring resource tests to better class
